// types/vendor.ts
export interface Vendor {
  id: string
  name: string
  slug: string
  logo: string
  coverImage: string
  description: string
  productCount: number
  rating: number
  address?: string
  phone?: string
  email?: string
  website?: string
  joinedDate?: string
  reviewCount?: number
  school?: string
  category?: string
}